+++ horovod/common/timeline.h
@@ -21,6 +21,7 @@
 #include <fstream>
 #include <iostream>
 #include <mutex>
+#include <thread>
 #include <unordered_map>
 #include <vector>
 
@@ -46,20 +47,35 @@ struct TimelineRecord {
 
 class TimelineWriter {
 public:
-  void Initialize(std::string file_name);
+  void Initialize(std::string file_name,
+                  std::chrono::steady_clock::time_point start_time_);
+  void Shutdown();
   inline bool IsHealthy() const { return healthy_; }
+  inline bool Active() const { return active_; }
   void EnqueueWriteEvent(const std::string& tensor_name, char phase,
                          const std::string& op_name, const std::string& args,
                          long ts_micros);
   void EnqueueWriteMarker(const std::string& name, long ts_micros);
+  void SetPendingTimelineFile(std::string filename);
+  TimelineWriter();
 
 private:
   void DoWriteEvent(const TimelineRecord& r);
   void DoWriteMarker(const TimelineRecord& r);
   void WriterLoop();
+  void WriteAtFileStart();
+  std::string PendingTimelineFile();
+  void SetTimelineFile(std::string filename);
 
-  // Are we healthy?
+  // Are we healthy?  Queue no longer accepts new work items and stops draining
+  // immediately when false.
   std::atomic_bool healthy_{false};
+
+  // Similar to healthy, but allows queue to be drained before closing when set
+  // to false.
+  std::atomic_bool active_{false};
+
+  bool pending_status_;
 
   // Timeline file.
   std::ofstream file_;
@@ -72,6 +88,15 @@ private:
   // Mapping of tensor names to indexes. It is used to reduce size of the
   // timeline file.
   std::unordered_map<std::string, int> tensor_table_;
+
+  std::thread writer_thread_;
+  std::string cur_filename_;
+  std::string new_pending_filename_;
+  bool is_new_file_;
+  // stores actual wall clock when horovod timeline was initialized
+  long long start_time_since_epoch_utc_micros_;
+  // mutex that protects timeline writer state
+  std::recursive_mutex writer_mutex_;
 };
 
 enum TimelineState { UNKNOWN, NEGOTIATING, TOP_LEVEL, ACTIVITY };
@@ -81,6 +106,7 @@ enum TimelineState { UNKNOWN, NEGOTIATING, TOP_LEVEL, ACTIVITY };
 class Timeline {
 public:
   void Initialize(std::string file_name, unsigned int horovod_size);
+  void Shutdown();
   inline bool Initialized() const { return initialized_; }
   void NegotiateStart(const std::string& tensor_name,
                       Request::RequestType request_type);
@@ -96,6 +122,7 @@ public:
   void ActivityEnd(const std::string& tensor_name);
   void End(const std::string& tensor_name, std::shared_ptr<Tensor> tensor);
   void MarkCycleStart();
+  void SetPendingTimelineFile(std::string filename);
 
 private:
   long TimeSinceStartMicros() const;