+++ horovod/keras/__init__.py
@@ -19,6 +19,7 @@ import keras.backend as K
 from horovod.tensorflow import init
 from horovod.tensorflow import shutdown
 from horovod.tensorflow import size
+from horovod.tensorflow import is_initialized, start_timeline, stop_timeline
 from horovod.tensorflow import local_size
 from horovod.tensorflow import rank
 from horovod.tensorflow import local_rank