+++ horovod/torch/__init__.py
@@ -32,7 +32,8 @@ from horovod.torch.mpi_ops import broadcast, broadcast_async, broadcast_, broadc
 from horovod.torch.mpi_ops import alltoall, alltoall_async
 from horovod.torch.mpi_ops import join
 from horovod.torch.mpi_ops import poll, synchronize
-from horovod.torch.mpi_ops import init, shutdown, is_initialized
+from horovod.torch.mpi_ops import init, shutdown
+from horovod.torch.mpi_ops import is_initialized, start_timeline, stop_timeline
 from horovod.torch.mpi_ops import size, local_size, rank, local_rank
 from horovod.torch.mpi_ops import mpi_threads_supported, mpi_enabled, mpi_built
 from horovod.torch.mpi_ops import gloo_enabled, gloo_built